End Game Screen and Bugfixes
Added in player stats which track how much damage was dealt and how many kills were achieved. These now appear at the End of Game screen. I Plan to differentiate damage between damage to others and damage to self later. I also discovered the cause of that weird bug where sometimes it wouldn't let me click on a weapon which had spawned on top of a player until I moved it off of the player. I need to remember to Instantiate copies of the prefab weapons rather than the weapons themselves. The Game Now feels like it has a satisfying ending.