Character Color Selection
Now allow multiples of the same character to be selected but forces them to choose different colors. For some reason this reintrosduced the bug with not being able to click on the weapon over the character collider.

Also changed spawns so that player one always spawns in spawn1; player 2 always spawns in spawn2; etc.