Fixed explosions
Fixed the explosions so that now they appear. Also, pinpointed the bug. It seems that if the character moved in between spawning the weapon and throwing it, then the weapon won't throw. Somehow the colliders get Screwed Up.

Added A main menu which should allow for a level select should we choose to do that.